Make action revisions viewable. Setup debug mode hook.
This commit adds the ability to look at prior action revisions. It
also prevents the focus from adjusting to the head when history
changes if you are not looking at the head node of the tree.

Additionally, it configures the devtools to use the new `debug: true`
hook for the next release of Microcosm.